Update patch set 2
Patch Set 2:

Since it is a worthwhile question, performance.

First reference is https://wiki.ubuntu.com/DashAsBinSh
Alas this is yet another example of sources of information were plentiful when it was written and now finding actual numbers is difficult.  This though does provide the tidbit of the change happened with Ubuntu 6.10.

That then allows finding this: http://www.linuxforums.org/articles/ubuntu-6-10-review_251.html
Apparently switching from bash to dash accounts for about a 50% reduction in boot time.  Figuring boot times are about 50% daemons and 50% scripting that gives an estimate of /bin/sh being as much as four times faster than /bin/bash.

Some more looking found someone doing something approximating actual benchmarking: https://unix.stackexchange.com/questions/148035/is-dash-or-some-other-shell-faster-than-bash
They're seeing /bin/sh being faster by a factor of 2-4 depending on which shell is /bin/sh.  Perhaps this script is a small part of the LineageOS build process, but a factor of 4 is nothing to sneeze at even if in what amounts to a shell fragment.



In other news, apparently SunOS^WSolaris had a pure Bourne shell as /bin/sh until very recently.  The switchover was 2010 apparently:
https://blogs.oracle.com/otn/new-shell-in-oracle-solaris-11

Unfortunately this means there are still people on SunOS^WSolaris 5.10^W2.10^W10 who want scripts to run on their machines even now:
https://github.com/rust-lang/rust/issues/25845

Perhaps what was broken into a second commit isn't critical, but it is a simple change to ensure compatibility with those systems.

Patch-set: 2