Update patch set 4
Patch Set 4:

Not sure why the h918 has anything to do with the h990, which is a completely tree. If that's you intent you should be using that one. Also, no other device tree has this, so this is not needed.

Patch-set: 4
Reviewer: Eric Meddaugh <14737@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>